fix: condition code not working on some website
