+++ src/workflowEngine/blocksHandler/handlerJavascriptCode.js
@@ -7,10 +7,10 @@ import {
   jsContentHandlerEval,
 } from '../utils/javascriptBlockUtil';
 import {
+  waitTabLoaded,
   messageSandbox,
   automaRefDataStr,
-  waitTabLoaded,
-  sendDebugCommand,
+  checkCSPAndInject,
 } from '../helper';
 
 const nanoid = customAlphabet('1234567890abcdef', 5);
@@ -70,39 +70,8 @@ async function executeInWebpage(args, target, worker) {
     return result;
   }
 
-  const [isBlockedByCSP] = await browser.scripting.executeScript({
-    target,
-    func: () => {
-      return new Promise((resolve) => {
-        const eventListener = ({ srcElement }) => {
-          if (!srcElement || srcElement.id !== 'automa-csp') return;
-          srcElement.remove();
-          resolve(true);
-        };
-        document.addEventListener('securitypolicyviolation', eventListener);
-        const script = document.createElement('script');
-        script.id = 'automa-csp';
-        script.innerText = 'console.log("...")';
-
-        setTimeout(() => {
-          document.removeEventListener(
-            'securitypolicyviolation',
-            eventListener
-          );
-          script.remove();
-          resolve(false);
-        }, 500);
-
-        document.body.appendChild(script);
-      });
-    },
-    world: 'MAIN',
-  });
-
-  if (isBlockedByCSP.result) {
-    await new Promise((resolve) => {
-      chrome.debugger.attach({ tabId: target.tabId }, '1.3', resolve);
-    });
+  const { debugMode } = worker.engine.workflow.settings;
+  const cspResult = await checkCSPAndInject({ target, debugMode }, () => {
     const { 0: blockData, 1: preloadScripts, 3: varName } = args;
     const automaScript = getAutomaScript(
       varName,
@@ -116,30 +85,9 @@ async function executeInWebpage(args, target, worker) {
       preloadScripts,
     });
 
-    const execResult = await sendDebugCommand(
-      target.tabId,
-      'Runtime.evaluate',
-      {
-        expression: jsCode,
-        userGesture: true,
-        awaitPromise: true,
-        returnByValue: true,
-      }
-    );
-
-    const { debugMode } = worker.engine.workflow.settings;
-    if (!debugMode) await chrome.debugger.detach({ tabId: target.tabId });
-
-    if (!execResult || !execResult.result) {
-      throw new Error('Unable execute code');
-    }
-
-    if (execResult.result.subtype === 'error') {
-      throw new Error(execResult.description);
-    }
-
-    return execResult.result.value || null;
-  }
+    return jsCode;
+  });
+  if (cspResult.isBlocked) return cspResult.value;
 
   const [{ result }] = await browser.scripting.executeScript({
     args,
